task 6
delete task functionality
