+++ src/store/getters.js
@@ -1,0 +1,5 @@
+export default {
+  allTodos(state) {
+    return state.todos;
+  }
+};
