+++ src/store/index.js
@@ -1,34 +1,13 @@
 import { createStore } from 'vuex';
+import mutations from '@/store/mutations';
+import actions from '@/store/actions';
+import getters from '@/store/getters';
 
-const state = {
-  todos: []
-};
+const state = { todos: [] };
 
 export default createStore({
-  state: state,
-  mutations: {
-    addTodo(state, title) {
-      state.todos.push({ id: Date.now(), title, completed: false });
-    },
-    toggleComplete(state, id) {
-      const todo = state.todos.find((todo) => todo.id === id);
-      if (todo) {
-        todo.completed = !todo.completed;
-      }
-    }
-  },
-  actions: {
-    addTodo({ commit }, title) {
-      commit('addTodo', title);
-    },
-    toggleComplete({ commit }, id) {
-      commit('toggleComplete', id);
-    }
-  },
-  getters: {
-    allTodos(state) {
-      return state.todos;
-    }
-  },
-  plugins: []
+  state,
+  mutations,
+  actions,
+  getters
 });
